+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Throw an exception instead of warning when both `localPath` and `version` are specified
 - Set `-Djdk.module.illegalAccess.silent=true` flag by default to mute `WARNING: An illegal reflective access operation has occurred`
 - `instrumentCode` task – incremental instrumentation [#459](../../issues/459)
+- Release `plugin.xml` file after reading it [IDEA-291836](https://youtrack.jetbrains.com/issue/IDEA-291836)
 
 ## 1.5.3
 - Updated dependencies marked as vulnerable